Add measurements ("measurements" table) to a SumavaDB-like spatiotemporal database

Add measurements ("measurements" table) to a SumavaDB-like spatiotemporal database

Usage

add_measurements_into_ssdb(
  conn,
  measurements,
  new = FALSE,
  append = !new,
  sql_files = system.file("sql/create_or_replace_functions.sql", package = "GSODDB"),
  warn = TRUE,
  err = !warn,
  nn_measurements = default_nn_measurements(),
  ...
)

Arguments

conn

database (preferably PostgreSQL) connection. See output of dbConnect.

measurements

measurements table, data frame or an object that can be coerced to a data frame. See output of sumava_snow_dataset

new

logical. If new==TRUE database tables are created or replaced. If new==FALSE tables are modidied according to append. Default is FALSE.

append

logical. If append==FALSE database tables are created or replaced. If append==FALSE data and metedata are appended as new rows of the database tables. Default is append=TRUE.

sql_files

file names containing database initialization SQL queries. They are used in case append==FALSE or new=TRUE.

warn

logical argument if the column names of measurements are different from nn_measurements functions displays a warning message. Default is TRUE.

err

logical argument if the column names of measurements are different from nn_measurements functions returns an error. Default is !warn.

nn_measurements

names of the columns in the "measurements" table. See the vignette.

...

further arguments (currently not used)

Examples

if (FALSE) { # \dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
val <- sumava_snow_dataset()
conn = dbConnect(PostgreSQL(), dbname = dbname)

add_locations_into_ssdb(conn,val$locations,new=TRUE)
add_measurement_types_into_ssdb(conn,val$measurement_types)
add_measurements_into_ssdb(conn,val$measurements)

### Add weather station from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_locations_into_ssdb(conn,val_smet$locations,append=TRUE)
add_measurement_type_into_ssdb(conn,val_smet$measurement_type,append=TRUE)
add_measurements_into_ssdb(conn,val_smet$measurements,append=TRUE)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- dbReadTable(conn,"measurements")
dbDisconnect(conn)

} # }