Union of one or more Weather Datasets from (UDER DEVELOPMENT)

Union of one or more Weather Datasets from (UDER DEVELOPMENT)

Usage

datasets_union(
  x,
  ...,
  warn = FALSE,
  err = !warn,
  nn_measurement_types = default_nn_measurement_types(),
  nn_locations = default_nn_locations(),
  nn_measurements = default_nn_measurements()
)

Arguments

x

dataset, e.g. an object returned by gsodr_dataset

...

further arguments, e.g other object like x.

warn

logical argument if any elements of x,... are not dataset-like objects, function displays a warning message. Default is TRUE.

err

logical argument if any elements of x,... are not dataset-like objects, function returns an error. Default is !warn.

nn_measurement_types

see add_measurement_types_into_ssdb.

nn_locations

see add_locations_into_ssdb.

nn_measurements

see add_measurements_into_ssdb.

Examples

 library(readr)
 x <- system.file("west_africa_gsod_extdata/gsod_bf_v2.csv",package="GSODDB") %>% 
 read_csv() %>% gsodr_dataset()
#> Rows: 3254 Columns: 47
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (9): STNID, NAME, CTRY, COUNTRY_NAME, ISO2C, ISO3C, MAX_ATTRIBUTES, MI...
#> dbl  (35): LATITUDE, LONGITUDE, ELEVATION, BEGIN, END, YEAR, MONTH, DAY, YDA...
#> lgl   (2): STATE, SNDP
#> date  (1): YEARMODA
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Joining with `by = join_by(location_code0, time, variable)`
 y <- system.file("west_africa_gsod_extdata/gsod_ml.csv",package="GSODDB") %>% 
 read_csv() %>% gsodr_dataset()
#> Rows: 4745 Columns: 47
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (9): STNID, NAME, CTRY, COUNTRY_NAME, ISO2C, ISO3C, MAX_ATTRIBUTES, MI...
#> dbl  (35): LATITUDE, LONGITUDE, ELEVATION, BEGIN, END, YEAR, MONTH, DAY, YDA...
#> lgl   (2): STATE, SNDP
#> date  (1): YEARMODA
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Joining with `by = join_by(location_code0, time, variable)`
 z <- system.file("west_africa_gsod_extdata/gsod_sn.csv",package="GSODDB") %>% 
 read_csv() %>% gsodr_dataset()
#> Rows: 4272 Columns: 47
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (9): STNID, NAME, CTRY, COUNTRY_NAME, ISO2C, ISO3C, MAX_ATTRIBUTES, MI...
#> dbl  (35): LATITUDE, LONGITUDE, ELEVATION, BEGIN, END, YEAR, MONTH, DAY, YDA...
#> lgl   (2): STATE, SNDP
#> date  (1): YEARMODA
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Joining with `by = join_by(location_code0, time, variable)`

 out <- datasets_union(x,y,z)
 
 outl <- datasets_union(list(x,y,z))
 
 is_dataset(out)
#> [1] TRUE
 is_dataset(outl)
#> [1] TRUE