Create GSOD Weather Dataset from (UDER DEVELOPMENT)

Create GSOD Weather Dataset from (UDER DEVELOPMENT)

Usage

gsodr_dataset(
  x = NULL,
  ...,
  description_url = "https://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt",
  location_code0 = "GSOD_%s",
  variable_code0 = "GSODR_%s",
  location_url = "default",
  use_limitations = "default",
  location_source = "GSOD",
  crs = 4326,
  measurement_types = "default"
)

Arguments

x

data frame returned from a call of get_GSOD. Default is NULL.

...

arguments for get_GSOD

description_url, location_url, location_source, use_limitations

attributes for locations table. See locations table and default.

location_code0, variable_code0

attributes for locations and measurement_types tables, as above. If it is one-element,it may contain a sub-string "%s" that will be replaced with the GSOD station identification code (STNID).

crs

Coordinate refence system. Default is 4326 (epsg).

measurement_types

data frame from a CSV file containing table with measuremet type attributes . It must contain an additional column has_flag .See measurement_types table and default.

Examples

 library(readr)
 x <- system.file("west_africa_gsod_extdata/gsod_bf.csv"
  ,package="GSODDB") %>% read_csv()
#> Rows: 2345 Columns: 47
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (9): STNID, NAME, CTRY, COUNTRY_NAME, ISO2C, ISO3C, MAX_ATTRIBUTES, MI...
#> dbl  (33): LATITUDE, LONGITUDE, ELEVATION, BEGIN, END, YEAR, MONTH, DAY, YDA...
#> lgl   (4): STATE, STP, GUST, SNDP
#> date  (1): YEARMODA
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
 out <- gsodr_dataset(x=x)
#> Joining with `by = join_by(location_code0, time, variable)`
 
 if (FALSE) { # \dontrun{
 t <- system.time(out <- gsodr_dataset(years=2020,country="IT"))
 } # }