Create MeteoTrentino Weather Dataset from a directory of SMET files downloaded by the Trentino Open Data Portal

Create MeteoTrentino Weather Dataset from a directory of SMET files downloaded by the Trentino Open Data Portal

Usage

meteotrentino_smet_dataset(
  data_dir = system.file("smet_extdata/meteotrentino", package = "GSODDB"),
  smet_files = NULL,
  crs = 4326,
  country_name = "Italy",
  country_code_iso_3166_1 = "IT",
  country_code_iso_3166_2 = "IT-TN",
  location_source = "MeteoTrentino",
  location_url = "http://storico.meteotrentino.it/web.htm?ppbm=%s&rs&1&df",
  measurement_types = read.table(system.file("smet_extdata/smet_variables.csv", package =
    "GSODDB"), header = TRUE, sep = ",", stringsAsFactors = FALSE),
  use_limitations = "https://dati.trentino.it/dataset/dati-recenti-delle-stazioni-meteo",
  location_description =
    "https://dati.trentino.it/dataset/dati-recenti-delle-stazioni-meteo",
  ...
)

Arguments

data_dir

directory containing original CHMI and DWD snow data

smet_files

SMET files used contained in the data_dir directory. Full path to SMET files if data_dir is NULL.

crs

Coordinate refence system. Default is 4326 (epsg).

country_name, country_code_iso_3166_1, country_code_iso_3166_2, location_source, use_limitations, location_description

attributes for locations table. See locations table and default.

location_url

attribute for locations table, as above. If it is one-element,it may contain a sub-string "%s" that will be replaced with the SMET station identification code.

measurement_types

CSV file containing table with measuremet type attributes . See measurement_types table and default.

...

further arguments

Examples

 data_dir <- system.file("snowdb",package="GSODDB")
 if (FALSE) { # \dontrun{
 out <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
 } # }