Create Sumava Snow Dataset from DWD and CHMI sources

Create Sumava Snow Dataset from DWD and CHMI sources

Usage

sumava_snow_dataset(
  data_dir = system.file("snowdb", package = "GSODDB"),
  remove_multipoints = TRUE,
  snap_distance = 5
)

Arguments

data_dir

directory containing original CHMI and DWD snow data

remove_multipoints

logical. It splits station and related time series if a station has resulted in different (even if near) point sites in different time periods (e.g. some CHMI stations). DEfault is TURE.

snap_distance

maximum distance expressed in meters between points for snapping. Default is 5.

Examples

 data_dir <- system.file("snowdb",package="GSODDB")
 if (FALSE) { # \dontrun{
 out <- sumava_snow_dataset()
 is_dataset(out)
 } # }